+++ src/config.js
@@ -9,10 +9,21 @@ export default {
     'https://chromewebstore.google.com/detail/bandcamp-to-discogs/hipnkehalkffbdjnbbeoefmoondaciok',
   firefox_extension_url:
     'https://addons.mozilla.org/en-US/firefox/addon/bandcamp-to-discogs/',
-  discogs_search_artist_url:
-    'https://www.discogs.com/search?q={artist}&type=artist',
-  discogs_search_release_url:
-    'https://www.discogs.com/search?q={artist}+{release}&type=release',
+  bandcamp: {
+    search: {
+      all: 'https://bandcamp.com/search?q={query}',
+      artist: 'https://bandcamp.com/search?q={query}&item_type=b',
+      release: 'https://bandcamp.com/search?q={query}&item_type=a'
+    }
+  },
+  discogs: {
+    search: {
+      all: 'https://www.discogs.com/search/?q={query}',
+      artist: 'https://www.discogs.com/search?q={artist}&type=artist',
+      release:
+        'https://www.discogs.com/search?q={artist}+{release}&type=release'
+    }
+  },
   text: {
     // Submission notes text
     notes: